WIP: Working on connecting test_score.html to postgres database
test_score.html is a mini form that I created to be used as an example during our development. If we can get the entire pipeline of this form working (get the information for a form, POST, recieve the request via Django, process and post to Postgres), then we can extend this to the other two forms.

Currently, the form sends a POST request with the information from test_score's form. I am running into issues submitting the form and going to the submitted.html page. We also need to verify that the information is being posted and saved to the database.